generalize templating into utility function + enum
