+++ rust/routee-compass-codegen/src/generator/mod.rs
@@ -3,7 +3,6 @@ pub mod constraint;
 mod error;
 pub mod input_plugin;
 pub mod output_plugin;
-pub mod traversal;
 pub mod util;
 
 pub use component_type::CodegenComponentType;
